refactor: move v1 REST Controller into Contracts namespace
- Relocate abstract Controller to Rest\V1\Contracts
- Import the contract in Rest\V1\Provider and mirror the test path
